% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reason_code_lm.R
\name{reason_code.lm}
\alias{reason_code.lm}
\title{Reason Code for a linear model}
\usage{
\method{reason_code}{lm}(model, newdata, ...)
}
\arguments{
\item{model}{a model object}

\item{newdata}{data for prediction of the same format as was originally fit}

\item{...}{further arguments for other methods}
}
\value{
a data.frame containing reasons
}
\description{
blah blah blah
}
\examples{
lm1 <- lm(mpg ~ cyl + wt, data = mtcars[-(1:2),])
reason_code(lm1, mtcars[1:2,])
}
\seealso{
\code{\link{reason_code}}
}
