% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reason_code.R
\name{reason_code}
\alias{reason_code}
\title{Generic reason_code method}
\usage{
reason_code(model, newdata, ...)
}
\arguments{
\item{model}{a model object}

\item{newdata}{data for prediction of the same format as was originally fit}

\item{...}{further arguments for other methods}
}
\value{
a data.frame containing reasons
}
\description{
blah blah blah
}
\seealso{
\code{\link{reason_code.lm}}
}
